Changed ricky name to index.js
